Support UDP sockets sharing a Port
Upon receiving a command from a given IP Address and Port number, Sennheiser
devices (at least those using MCP over UDP) will send replies to the given IP
Address, but not to the given Port number, preferring instead to send it to a
specific Port.

Thus, in order to support monitoring more than one MCP device at the same time,
we need to support sockets sharing a Port. We also need to determine from which
device we're receiving a message, else replies end up being routed incorrectly
upon receipt.

Further note: we need to disconnect the socket on config change, as otherwise
further data will be received on the old socket (which is still active, but not
referenced anywhere), and not on the new one.