Set initial timestamp to `0`
We hold the "last battery" status for 30 mins, and use this property to
determine when we last saw a RX device. Except it was being initialized to
`current time - 1 minute`.

Thus: if a RX connected without an active TX, then the "last battery" status
was being sent to the UI as if the TX had been turned off just moments ago.