Move call to init logging out of `config.config()`
Previously it was called everytime the config changed which, due to logging
handlers being created (and the old ones not being removed), was causing
all subsequent log entries to be written to stdout and the log file more than
once.

Now, it's only init'd once, and the duplications are no more.