Tentatively add SLX-D support
Don't actually have a device to test it with, so this is based on the
documentation.

The DCID_MODEL value comes from the DCID of Shure Update Utility (v2.4.8)
Annoyingly, they appear to use the same DCID_MODEL value for both the
single-channel SLXD4 and the dual-channel SLXD4D models.

I don't know the DCID names for the four other devices, as those came out
more recently.