Build address for QR based on actual connection
If `local_url` is not set in the config file, then the default address of the
default interface is used instead.

On some devices, the "default" interface is the loopback device, causing the
address to be "127.0.0.1". This is not helpful for use generating a QR code,
so use the address of the requesting client instead.

(Of course, if the client is connected via loopback, then there'll be no
difference to what was provided before, but that's another kettle of fish.)